requirements.txt
# requirements.txt

# Web framework and server
fastapi==0.111.0
uvicorn[standard]==0.29.0
jinja2==3.1.4
python-multipart==0.0.9

# ML and data handling
# Use the version you confirmed works with your model
tensorflow==2.19.0
numpy==1.26.4
Pillow==11.2.1
